Enhance job retrieval
Jobs are used as a communication pipe between the main web process and
the background processing (celery).
Jobs are initiated by the web app and updated by the celery process.

Sometimes the transaction of the request has not been commited yet, the
job is not persisted in database when the background process starts (it
starts reading the redis database informations)